Refcatored the sections folder to utilize React rather then js based React
