+++ src/components/PortfolioContainer.js
@@ -1,10 +1,10 @@
 import React, { useState, useEffect, useRef, useMemo, useCallback } from 'react';
-import IntroSection from './sections/IntroSection';
-import SkillsSection from './sections/SkillsSection';
-import ExperienceSection from './sections/ExperienceSection';
-import ProjectsSection from './sections/ProjectsSection';
-import EducationSection from './sections/EducationSection';
-import ContactSection from './sections/ContactSection';
+import IntroSection from './sections/IntroSection.jsx';
+import SkillsSection from './sections/SkillsSection.jsx';
+import ExperienceSection from './sections/ExperienceSection.jsx';
+import ProjectsSection from './sections/ProjectSection.jsx';
+import EducationSection from './sections/EducationSection.jsx';
+import ContactSection from './sections/ContactSection.jsx';
 import './PortfolioContainer.css';
 
 const PortfolioContainer = ({ onScrollChange, deviceType }) => {
